[sp-36] user data editing
-Moved editing form to seperate file
-Added success toast
